can fetch data progress and enrollment n complete

+++ lib/app/modules/vdo_page/controllers/vdo_page_controller.dart
@@ -156,7 +156,7 @@ class VdoPageController extends GetxController {
       if (element.videoId == currentVideoId) {
         var videoProgress = currentEnrollment.progress
             .firstWhere((element) => element.videoId == currentVideoId);
-        videoProgress.progressPercentage = percentage;
+        videoProgress.progressPercentage = percentage.toInt();
         videoProgress.lastViewedTimestamp = DateTime.now();
         break;
       }
